DJANGO/folder/project/RollVision/requirements.txt
Django>=4.2.0,<5.0
opencv-contrib-python>=4.8.0
Pillow>=10.0.0
reportlab>=4.0.0
numpy>=1.24.0
python-decouple>=3.8
